requirements.txt
Appium-Python-Client==0.26
selenium==3.11.0
